+++ src/channel.ts
@@ -1211,7 +1211,7 @@ export class Channel<StreamChatGenerics extends ExtendableGenerics = DefaultGene
       }),
     };
 
-    this.getClient().polls.hydratePollCache(messageSet.messages, true);
+    this.getClient().polls.hydratePollCache(state.messages, true);
 
     const areCapabilitiesChanged =
       [...(state.channel.own_capabilities || [])].sort().join() !==